Add support for SDI output of SCTE-104 messages, when SCTE-35 in stream
Update the decklink SDI output interface to support receiving SCTE-35
data streams, converting them to SCTE-104, and putting them out as
VANC.

Note that we had to add a new AVPacket sidechannel item for
the original PTS coming from the demux.  This is because all the
PTS values get recomputed before sending to the output, and we
need the original PTS in order to compute times for pre-roll
events.  Unlike most protocols where the PTS represents when the
packet should hit the output, in the case of SCTE-35/104 the PTS
represents some time in the future after that.  Because of this
we need to be able to subtract the specified PTS from the PTS
of the packet (i.e. based on the program PCR) in order to determine
what the actual time should be.

The functionality can be tested with:

./ffmpeg -i ~/test.ts -map 0 -f decklink -vcodec v210 'DeckLink Duo (1)'

We require 10-bit video out in order to output SCTE-104.  This is a
requirement of SMPTE ST2010-2008 (although we should probably warn the
user if they try to output SCTE-104 when in 8-bit mode that it isn't
going to work).