Support encoding of Active Format Description (AFD) in libx264
If AFD side data is present, include it in an H.264 SEI payload when
encoding with libx264.

This is done in the same manner that we currently handle A53 closed
captions (where the business logic for constructing the SEI is in
libavcodec/utils.c), so it should be portable to the other encoder
types (i.e. videotoolbox, etc).

Updated to reflect feedback from Derek Buitenhuis <derek.buitenhuis@gmail.com>
and Aaron Levinson <alevinsn_dev@levland.net>.

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>