Control whether SMPTE 2038 is mapped based on environment variable
Because we don't know in advance whether there will be SMPTE 2038
in the transport stream, make the auto-configuration logic forward
it if present, but only if an environment variable says to.

This avoids the controller having to probe the TS input and then
explicitly create a "-map" argument.  This seems a bit redundant
in the case of SMPTE 2038 since it would have been selected
automatically because the decklink codec for data is 2038 by
default, but this becomes more important when other streams such
as SCTE-35 are involved.