Add bitstream filter to convert SCTE-35 to SCTE-104
Introduce a bitstream filter which converts from the SCTE-35 format
to SCTE-104.  The primary use case for this will be MPEG transport
streams which contain SCTE-35, where the goal is to output the
data as SCTE-104 over SDI (i.e. through the decklink output).

Note that this change introduces a new piece of side data to
preserve the original timestamp.  This is required in order for
the bitstream filter to compute preroll times, where the PTS
is embedded in the SCTE-35 packet.  To compute these deltas,
we need to know what the original timestamp was when the data
arrived in the demux.

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>