decklink: Add support for multiple streams of audio output
Provide the ability to output multiple streams of audio output
over the Decklink SDI interface, interleaving the samples from
each stream.

This functionality assumes either S16LE PCM samples at 48 KHz or
AC-3 packets (which get encapsulated in S337).  The implementation
is a bit naive in that it assumes the sample sizes across the
streams will be the same, and doesn't handle splitting sample payloads
across the timing window.  Hence it's possible that all streams
except the first could be -30ms behind schedule.  We would need to
implement a proper timing model to align the streams precisely.

It also keys the timing off the first stream, so if the first stream
stops delivering audio we will lose audio on all channels.

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>