Allow output of 720x480 and 704x480 over Decklink SDI
The Decklink SDI output for standard definition NTSC is fixed at
720x486.  However 720x480 and 704x480 are both common resolutions
found in the broadcast space.  Put in a hook to invoke the pad
filter to bring those resolutions up to 720x486.

Note that for 704 width we use the pad filter to add eight pixels
to each side as that seems to be the standard convention, at least
according to Video Demystified (Keith Jack) as well as it being
what the Sencore MRD-4400 does.  For the 480 height, we pad by 4/2
instead of 3/3 to properly preserve the field dominance.

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>