Don't stall the output if a SCTE-35 packet is encountered
The libavformat mux generally expects each stream to continuously
provide packets, which is fine for audio and video, but not a
reasonable assumption for SCTE-35 packets which can occur very
infrequently.

Add an exception to the mux to not wait for more packets.  Note
that we add both SCTE-35 and SCTE-104 because when the mux is
initialized the demux hands us a stream with the SCTE-35 codec
type, which the decklink modules inserts a BSF to convert to
SCTE-104 in write_header().  After the stream is running, any
actual packets that hit the mux will already have gone through the
BSF and thus the codec will be SCTE-104.

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>