Support passing field order through entire pipeline
In order to support properly selecting interlaced formats on the
decklink output, we need to be able to propagate the field order
all the way from the decoder, through anything in the filter
pipeline, out to the encoder.  This has to happen as a property
of the filter links rather than AVFrame because there is no
guarantee that any frames have gone through filters prior to
creating the AVFormat output (in this case, the decklink output
device).

Add an additional property to the AVFilterLinks, and make sure it
gets carried through the entire pipeline so that it can be
available when doing the write_header() call for the decklink
output.