Force audio output to stereo for all streams
We don't expose in the controller configuration how audio streams
should be output (discrete, stereo pairs, passthrough), so for
now just force it to stereo.  It also looks like there might be
audio distortion in playing out 5.1 discrete mode, but this might
be a problem with the downstream SDI monitor.