Disable dts_delta_threshold
If there is a period where no data is received, and it lasts more
than dts_delta_threshold (defaults to 10s), then the PTS will be
set to the next pts from the last properly received PTS.  The
problem with this is the decklink output expects packets to be
scheduled relative to realtime, so all subsequent packets will
be sent to the mux with a PTS that is more than 10s late.  This
causes total loss of audio and video will only periodically output
frames (the Blackmagic docs say on each interval the earliest late
frame will be output and all earlier ones dropped).

Disable this logic, such that if the stream stops and then restarts
the PTS will continue at the new packet.