decklink: Redesign support for multi-channel audio output
Restructure the way we do audio output such that we now defer
scheduling to the hardware until 0.25 seconds before realtime.
This provides the opportunity for all streams to have delivered
audio by that point, and if not then that channel delivers silence.
It also avoids an edge case where previously we would wait until
the next packet on the first stream to schedule the previous packet
of audio, which can cause scheduling to be too late if we receive
large batches of packets from a single stream (as opposed to a
series of interleaved payloads from each stream).

Also rework the way we do the timing for interleaving to be accurate
down to a single sample across all audio pairs.  This addresses cases
where the various streams are not delivering blocks of raw PCM
samples aligned on the same PTS boundary (common when the different
audio streams are encoded with different codecs).  This also ensures
there is no loss of A/V sync in the case where there might be a gap in
audio (e.g. due to a lost audio packet), whereas previously we would
have just concatenated the samples after the last successful packet.

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>