Use updated API to ensure standard definition VANC is properly written
Because standard definition SDI video makes use of both the Luma and
Chroma channels for VANC, we have created a new libklvanc API which
automatically does "the right thing" for both HD and SD video formats.

Make use of that new API, which as a side-benefit means the caller
no longer has to do the final colorspace conversion since it now
happens in libklvanc.