Fix two bugs in generation of CEA-708 CDP messages
During regression testing there were two inconsistencies found in
creating of CEA-708 CDP messages:

1.  We were not setting the caption_service_active flag, which needs
to be set in addition to just indicating that the CC section is
present.

2.  The check for whether a CC triple is active was checking the
wrong bit, resulting in it saying all the triples were active
rather than just those which really have valid data.

Neither of these are known to cause issues with any specific piece
of downstream equipment, but it makes sense to properly conform
to the spec just in case.