Properly output 480i NTSC video that is encoded as TFF
The SDI standard requires that 480i video be output BFF, but it's
possible for the encoding source to encoded it as either
top-field-first or bottom-field-first.  In fact, top-field-first is
the more common approach with H.264 encoded video.

Automatically insert the vf_phase video filter into the pipeline
if we encounter 480i video that is TFF.

Note that this is about field dominance - the temporal order in
which fields within a frame are transmitted on an interlaced output.
This is *not* about the fields being inverted within the frame.
Hence in order to convert to BFF we have to sample two separate
frames, and recombine the bottom field from the previous frame
with the top field from the current frame.  This is why we use
the "phase" filter as opposed to the "fieldorder" filter (which
is what one might assume would be the intuitive choice).