Deal with possibility that streams are not all in same program
The existing heuristic ffmpeg employs to pick the "best" audio and
video to play doesn't take into consideration that the streams may
not be coming from the same TS program.  This can result in playback
problems, in particular if the streams are operating off of
different PCR clocks.  Also, it's highly unlikely that somebody
wants to watch the video from one program and listen to the audio
from another.

Take into account the program that streams belong to in selecting
which audio/video stream to play, and when multiple programs are
present then choose the first video and audio stream stream from
the first program in the PMT, rather than trying to choose the
"best" stream based on some quality metrics such as bitrate or
resolution.