Introduce video filter to inject MacCaption MCC files into streams
Provide the ability to take an MacCaption VANC MCC file, extract
the CEA-708 closed captions, and insert them into the video stream
as A53_CC frame side data.  This, when used in conjunction with an
encoder that supports A53_CC side data (e.g. libx264) allows for
splicing in the CEA-708 captioning data to an exising stream.

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>