Fix case where decklink card can get sick on output
This is essentially the same fix we did for VLC a couple of months
ago, where we need to proactively start the output with the previous
video mode and then stop it before attempting to use it for real
video output.  This addresses a case where the card gets sick if
the application previously received a SIGKILL and the allocated
buffers weren't properly freed.

This issue has been reproduced with the Decklink Duo 2 card when
using the 10.8.5 driver.  It is unknown whether other cards and/or
versions of the driver are effected.