Fix case where h->tc_dropframe might not be properly set
We were setting h->tc_dropframe to the value of the cnt_dropped_flag,
which is incorrect given it could be set but the value of
counting_type could be zero.

Introduce a temporary variable to represent the value of
cnt_dropped_flag, and use it to determine whether to set the
h->tc_dropframe flag to 1.