Fix PTS wraparound in audio
A temporary variable used for computing the audio window for
playout was defined as an int instead of an int64_t.  The result
was that you would have a negative PTS window when the counter
exceeded 31 bits.  This would result in no audio being scheduled
after this time.

Trivial to demonstrate by adding "-itsoffset 47580" as a command line
option.