Fix failure to play progressive video from OBE
If the pic_timing is not present in the stream (as is the case for
progressive streams coming out of libx265), treat the video as
progressive.  This occurs because the field_order is set to
unknown and the SEI parser never gets invoked, which is what would
normally set it to some specific mode.