Handle cases where a filter changes width/height
The filter management code was only looking at the original
width/height at the front of the pipeline, even if a filter in
the pipeline resulted in the values being changed.  In particular
the use of the fieldmerge filter doubles the height, but that new
height wasn't being reflected when determining whether to use the
subsequent vf_pad filter.

Track the updated width/height so it can be used in subsequent
filter decisions.