Don't add audio streams without sufficient info
If we don't receive any packets that correspond to an audio stream
in the PMT, we won't have enough info to make decisions to set up
the output (i.e. the number of channels and sample rate).  Exclude
the stream from the output rather than bailing out completely at
initialization.