Ensure frames produced by vf_fieldmerge are TFF
If the filter happens to receive a bottom frame first, the temporal
alignment is off for the filter.  This results in playout of the
fields in the order 2143 instead of 1234.

If the first field we ever receive is a bottom field, throw it away.
This ensures frame pairs received are temporarly adjacent.

The alternative would be to properly set the top_field_first flag
in the resulting frame based on whichever we received first.
However this results in then having to possibly use the vf_phase
filter to realign the resulting frames prior to output.  It could
also cause a difference between when the AVLinks are negotiated
compared to the actual frames being delivered down the pipeline.