Second attempt at PTS wraparound
The previous patch tracked PTS values and wrap state on a per-stream
basis, which doesn't work for streams which have very sparse
packets (e.g. SCTE-35).  In such cases, we might miss the
wraparound window and thus deliver packets from the demux that
weren't wrapped (causing the mux to abort).

Refactor the code to more closely follow the upstream model, which
does have the state tracking on a per-stream basis, but when a
wraparound trigger occurs, have it update the state for all streams
and not just the stream that hit the wraparound.

The delta for this patch is pretty ugly, so anyone looking too
closely at it might be worthwhile to combine it with the previous
patch before comparing against upstream.  This is particularly
true given the second patch reverts some previous changes which
make the resulting delta smaller compared to upstream.