Add video filter to remove 608/708 caption data from A53 side data
While the "vf_sidedata" allows for removal of the entire a53
side data block, a user might want to just strip out the 608 or
708 captions (to deal with problems in downstream decoders).

Introduce a filter which allows the user to specify which should
be deleted, replacing it with DTV padding.

See ANSI/CTA 708-E Sec 4.3 for the semantics of the cc_data structure.

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>