Fix segfault when referring to non-existent stream
The code used to find the "best" audio stream would look for a
video stream within the same program.  But if a video stream was
never found, we would just pick the first stream in the program.

The problem with this approach is we could have a TS that didn't
associate any streams with the program (i.e. an program with
no entries).  In this case we would dereference a NULL pointer by
attempting to refer to the first entry in a list that doesn't
contain any entries.

It's actually perfectly valid to pass -1 to av_find_best_stream(),
so if we don't find video then just do that.  Because video_idx
is initialized to -1, we can get rid of the intermediate variable
"related_stream" in the process.

This problem manifested with a very short TS where there was only
one PAT and one PMT, and the PMT actually came before the PAT.
The problem was worse because of the ts->auto_guess functionality
which discovers the PES packets and creates streams that aren't
associated with the program.  This resulted in three streams that
weren't tied to any program, and a program with zero streams.