Fix incorrect setup of stream map when PMT order doesn't match order in AVProgram
Because of the design of the probing behavior, there could be
cases where the PMT order doesn't match the order of the entries
in the AVProgram stream_index.  In these cases, referring to the
index based on the placement within the array causes a mismatch with
later routines which refer to the stream ID.

This resulted in cases, for example, where we created a streammap
that had a SCTE-35 stream being specified as the video stream.

Example of incorrect stream map construction:

Input #0, rtp, from 'rtp://@227.1.1.7:4007':
  Duration: N/A, start: 72133.228811, bitrate: N/A
  Program 1
    Stream #0:2: Audio: mp2, 48000 Hz, stereo, s16p, 256 kb/s
    Stream #0:1: Audio: mp2, 48000 Hz, stereo, s16p, 256 kb/s
    Stream #0:3: Data: scte_35
    Stream #0:0: Video: h264 (Main), yuv420p(tv, bt709, progressive), 1280x720 [SAR 1:1 DAR 16:9], Closed Captions, 59.94 fps, 59.94 tbr, 90k tbn, 119.88 tbc
Stream mapping:
  Stream #0:3 -> #0:0 (? (?) -> v210 (native))
  Stream #0:0 -> #0:1 (h264 (native) -> pcm_s16le (native))
  Stream #0:1 -> #0:2 (mp2 (native) -> pcm_s16le (native))
  Stream #0:2 -> #0:3 (copy)

Note that in the above case, stream 0:3 (scte_35) is being
specified as video, and stream 0:2 (audio) is being specified
to do stream copy rather than being decoded.  Also, the stream
map has the H.264 stream being passed to an audio decoder.

Make sure any cases where we refer to streams are based on the
stream->index value rather than the position within the input_streams
array.