Fix logic to start start playback at earliest opportunity.
The driver needs to have enough frames prerolled before it can start
playback, but the current logic would effectively wait until there
were (frames_preroll + 2) before starting playback.  This was because
the code was looking for the frame number of the current frame to
be greater than the preroll count, when in fact it only had to be
equal to it, and a second logic error where it the math was comparing
against one greater than the actual preroll.

For example, if the preroll is 3 and the first PTS is 30, the playout
wouldn't start until after frame 34 was queued to the hardware.  In
reality, we could have started playout after frame 32 was queued
(since at that point frames 30, 31, and 32 would be pending).
.