decklink: Make the number of initial buffers to discard configurable
Rather than a hard-coded 30 frames, derive the number of frames to
drop taking into account the framerate, and make the value
configurable.

This was important for two reasons:

Video such as 1080p/59 would need to drop more frames than i59,
since the frames are arriving at 2X the rate.

Higher bitrate streams are more likely to create a backlog during
probing, and hence we want to drop more than 1 second of video
in certain cases to ensure the UDP FIFO is fully emptied at startup.

Allow the value to be hard-coded, and default it to 1.5 seconds,
which translates to 44 frames for 1080i59 and 90 frames for 1080p60.