Use DTS of video instead of PCR for SCTE-35 table timing
The existing code would use the current PCR at the time the SCTE-35
entry was encountered and use that to derive a PTS for the packet.
However this doesn't take into consideration that the PCR can be
hundreds of milliseconds ahead of the video stream.

Rather than using the PCR, find the video stream within the program
and use the most recent DTS as the PTS/DTS of the SCTE-35 trigger.

Without this patch, the SCTE-35 trigger may get output by the mux
much later than expected.