vf_burwriter: Support YUV 422P format in addition to RGB
Because the filter only operated in the RGB colorspace, we were
incurring a significant CPU cost in that we had to transform the
YUV to RGB, burn in the counters, and then transform RGB back to YUV.

Add a mode of operation to support YUV 422P, which is the native
format being put out by the v210dec module.  This allows us to
avoid any CSC when reading from SDI, burning in counters, and then
putting the result onto an SDI output.

On a Haswell platform, such an operation uses effectively 100% of
a single core, meaning it cannot keep up with realtime.  With the
change the CPU usage on the main thread is down to 40%.