vf_burwriter: Avoid copying the frame when burning in the counters
If the reference count on the input frame is one, that means we are
the only party who intendeds to ever receive the frame.  Thus we can
burn the counters directly into the existing frame rather than
making a copy for the output.

This optimization is cut/pasted from the code found in a bunch of
other video filters, so we are now working in a consistent manner.

For 1080p/60 video, this reduces CPU usage by 10% on the main
thread on Haswell.