burnreader: add support for 422p and detect no counters
Provide a variant of the reader which supports yuyv422p10, which
is the native format for both the v210 encoder and decoder.  This
avoids having to use libswscale to do any colorspace or chroma
conversion in cases where we are working with decklink capture or
output.

Also, add code to use the chroma region to detect whether counters
are present, as opposed to just looking at the luma and returning
garbage values in the case where no counters are present in the
video.  This takes advantage of the fact that we always set the
UV channel to 0x200 (10-bit) or 0x80 (8-bit), regardless of
whether the block is white or black.