Add support for being able to specify the 3G output Level (A/B)
Introduce a new decklink output argument which allows the user to
be able to specify whether to output 3G-SDI Level A.

Note that we store the value in the Blackmagic persistent configuration.
We do this because we don't want the output flipping back to Level B
when ffmpeg exits, which is the default behavior if you don't write
the setting to persistent storage.

Example usage:

./ffmpeg -i foo.ts -f decklink -vcodec v210 -use_3glevel_a true 'DeckLink Duo (1)'