Add support for specifying decklink cards by device number
Currently the code requires you to know the exact name of the card,
as specified by the card description and visible by running
"ffmpeg -sinks".  However, this name can be changed in the
Blackmagic Desktop Video application, and it's annoying to have
to tweak scripts when running on systems that might have different
model cards.

Introduce the ability to specify either the description of the card
(e.g. "DeckLink Duo (4)") or a value derived from the probing order
of the card (e.g. "decklink0").  Both values are available by
running the "ffmpeg -sinks" command, and both are supported in
order to be compatible with legacy behavior.

This should make it much easier for a user to run scripts across
different hosts and/or models of cards by simply specifying "decklink0"
rather than having to figure out the card description and tailor
the command line to match.