Don't touch 3g level setting if the user didn't specify the command line option
For the new use_3glevel_a option, have the default be "auto" which
results in ffmpeg not touching the value in the configuration.
This is more consistent with the principle of least surprise, as
not specifying the argument on the command line shouldn't result in
the application forcing the output configuration to Level B.

Hence, the uesr can specify "-use_3glevel_a true", "-use_3glevel_a false",
and that value will take effect, but if they don't specify either
ffmpeg will use whatever was configured globally in the Desktop
Video application.