Move LTN feedback socket code to libavformat
The original implementation for providing feedback to the LTN
controller was baked into the decklink output module, but that
prevents it from being used in other places.  Move the logic into
libavformat where it can be shared with other callers.

Worth noting that it's in libavformat instead of libavutil because
libavutil doesn't have any of the platform independent abstractions
for UDP sockets.