Stash parameters from input phase of pipeline for reusing during output setup
The ffmpeg construction of the avfiltergraph has an input stage
and an output stage.  We need to store the results of the input
stage and expose them to the construction of the output stage in
order to make intelligent decisions with regards to how the output
stage should be setup (i.e. what filters to install).