Move decklink specific filters to output stage
The original ltndecoder code always assumed the output was a
decklink device, and thus it was safe to insert some filters
specific to that use case anywhere in the pipeline.  However as
we look to use the ltndecoder binary for other use cases, it
makes more sense to move those filters to the end of the pipeline
and have them only instantiated if the output is actually a decklink
device.