Fix broken output scaler
When adding support for propagating the interlaced/tff status
along the pipeline, we were passing the fields off to the scaler
as parameters.  Turns up we made a couple of mistakes in the
parameters sent but never noticed because up to this point we
weren't using the output scaler.

Fix the output scaler so it starts up properly and does include
the correct interlaced status.