Don't invoke scaler for 704x480 to 720x480
We deal with this later in the pipeline where we insert padding
(which is the correct behavior), so we don't want to invoke the
scaler.

This is a by-product of having moved the decklink specific code
to the end of the pipeline, as previously it would have been
dealt with already and the video would already have been 720x480
before it hit the output scaler.