Add support for interlacing on output
Add a new "-interlace" option which corresponds to the "deinterlace"
command line argument.  The deinterlace argument will operate on
the input while the interlace argument will operate on the output.

Note we also make sure the framerate is 2x the real output since
the interlacing will effectively cut the framerate in half.