Add command line options to control deinterlacer quality vs. perf
For cases like the thumbnailer, we are not as sensitive to the
deinterlacing quality, so expose options to control the quality,
as well as the ability to dictate it for SD vs. HD streams.

This is to improve performance in cases where the original stream
is HD but the output is a very small image.