Fix case where SCTE-35 messages don't get delivered after PTS wrap
The previous change we made to rely on the video PTS didn't take
into account that ths AVStream cur_dts member is *after* any
wraparound logic has been performed.  Hence we end up incrementing
it twice (causing such packets to be delivered 26.5 hours late)

Truncate at 33-bits so that we are consistent with all other packets
that come out of the mpegts demux.