Add -xdtserror argument to handle "Non-monotonous DTS" errors
If we hit cases where ffmpeg reports ""Non-monotonous DTS in output
stream", which commonly happens on a sudden stream change, exit
out so that the transcoder can be restarted.

The logic itself in ffmpeg for that case is clearly broken, in
particular if the stream DTS jumps backward the process will
never recover.  For now just short-circuit the logic so we
can restart and get back to a known state.