Avoid null pointer dereference in avio_write()
If there is an I/O write error on the output, it's entirely
possible the application will still try to flush its output and/or
call av_write_trailer().  Such cases would result in a NULL
pointer dereference and segfault.

If the AVIOContext is NULL, just return (the function is void so
we cannot return an error condition).  Better to drop the data on
the floor than to crash.