Increase preroll if required based on source stream timing
There are cases where the nature of the stream might result in the
decklink output preroll needing to be higher than the user specified
value.  Keep track of the PTS delta for the incoming stream and
use that to increase the pre-roll if needed beyond what the user
specified.

This effectively changes the definition of the decklink preroll
value from "amount to preroll" to "minimum amount to preroll"
given we might make the actual preroll higher if needed to prevent
audio and/or video stuttering.