Make the UDP monitoring callback a global parameter
The UDP monitoring callback to notify the controller of events was
originally a decklink output specific parameter.  However this prevents
it from being used by applications that don't output over decklink,
as well as announcing events which might occur before the decklink
output is initialized.

Make it a global argument and set it up from ltndecoder.c, which
makes it common code and gets it out of the decklink module.