Remove constraint that audio be downmixed to stereo
We had a long-standing hack in the code to always downmix/upmix any
audio to stereo.  This was convenient in that it always yielded
predicatable output but prevents cases such as discrete 5.1 from
working.

Remove the hack.  The controller can be adjusted to specify "-ac 2"
if the intent is to force everything to be remixed to stereo.