Fix segfault on HTTP transfers that encounter I/O error
If there was a session interruption, the write would fail but then
the http client would still attempt a shutdown on the closed FD,
resulting in a NULL pointer dereference.

Since it's not clear whether enforcement should be in the ffurl
utility functions or it's the caller's responsibility to never
provide a NULL URLContext pointer, change the code to return an
error in the utility function (which might have other callers) as
well as not making the call in the first place during the shutdown.

Stack trace this addressed for reference: