mpegtsenc: properly support CBR muxrate with -copy_ts argument
When explicitly specifying the muxrate in the mpegtsenc muxer, the
code would assume that the first timestamp would always be zero.
This influences the addition of NULL padding to the stream.  However
if the user specified -copy_ts then the original timestamps are
being passed through, and thus the first PCR could be essentially
any value.

Without this patch, if the user specified a muxrate as well as
-copy_ts the mpegtsenc module would insert potentially many
gigabytes of NULL padding at the start of the stream.  Take note
of the actual first timestamp received in the first packet, and
use to determine the first PCR rather than assuming the first
timestamp will be zero.

This problem would manifest in a number of ways:

* If the user was doing UDP output without specifying a bitrate
  parameter there would be a network saturating burst on the
  outgoing interface.
* If the user was doing UDP output and did specify a bitrate, there
  would be potentially many gigabytes of NULL PID packets, and
  the input may stall long enough to cause the ffmpeg input to
  overflow/exit.
* If the user was writing the TS to a file it could end up creating
  a file that was many gigabytes before there were any valid
  audio/video packets.